Address codex review of PR #80
- Duplicate-grade lots: every lot keeps its sold floor and the first in
  canonical order carries the remainder — a later duplicate holding sold
  eggs no longer produces a spurious 422 when the aggregate target covers
  all sold eggs. Explicit aggregate-level 422 when it does not.
- Ledger notes truncate to BirdMovement.MaxNoteLength: a max-length reason
  no longer throws (500) when prefixed into the compensating movement note.
- Lock sweep uses a strict comparison — an entry exactly 7 farm-local days
  old keeps its final editable day; only strictly older entries lock.
- Loss-count sum validated as long in BOTH the adjust and record
  validators: three ints could overflow past a positive total and slip
  through the domain guard when no grade lines were present.
- AdjustedFromJson is plain text, not jsonb (provider-portability rule);
  the two migrations squashed into one clean AddDailyEntryAdjustments with
  the unambiguous lot backfill.
- Docs-sync in THIS PR: Help page (submit bullet, 7-day lock bullet, two
  Fixing-mistakes rows, three glossary rows) + GLOSSARY.md full lifecycle
  rewrite (Locked/ManagerAdjusted/Voided, lot linkage, sold-untouchable).
- New race test: parallel adjust vs void on the same base version —
  exactly one wins, Version delta 1, no blended lot state.

+++ src/Cluckwork.Application/Features/DailyEntries/AdjustDailyEntry/AdjustDailyEntryValidator.cs
@@ -14,8 +14,9 @@ public sealed class AdjustDailyEntryValidator : AbstractValidator<AdjustDailyEnt
         RuleFor(x => x.DirtyEggs).GreaterThanOrEqualTo(0);
         RuleFor(x => x.DiscardedEggs).GreaterThanOrEqualTo(0);
         RuleFor(x => x.MortalityCount).GreaterThanOrEqualTo(0);
+        // long accumulation — three ints can overflow past a positive total.
         RuleFor(x => x)
-            .Must(x => x.CrackedEggs + x.DirtyEggs + x.DiscardedEggs <= x.TotalEggs)
+            .Must(x => (long)x.CrackedEggs + x.DirtyEggs + x.DiscardedEggs <= x.TotalEggs)
             .WithName("Eggs")
             .WithMessage("Cracked + dirty + discarded cannot exceed total eggs.");
         // NotEmpty alone lets "   " through to the domain guard (500, not 400).
